Revert "core: add subchannel list LB commons utility function, refactor for WRR (#9875)"
This reverts commit 5a2adcc7c6ff32aeddcbd9eafdec194a3ac44186.
